+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 7.11.1 - 2026-08-04
+
+### Removed
+
+- Remove duplicate upgrade calls from publish workflow
+
+## 7.11.0 - 2026-08-04
+
+### Changed
+
+- Improve push and publish workflow
+- Rename »gg do upgrade dependencies« into »gg do upgrade deps«
+
 ## 7.10.0 - 2026-08-04
 
 ### Changed